backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
cryptography==42.0.8
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyjwt==2.8.0
pyotp==2.9.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.0
pytest==8.2.0

